# public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tournament API",
        "description": "API documentation for managing tournaments",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tournaments": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get all tournaments",
                "operationId": "8626d4f0c19b7c5d27dce357833b6b97",
                "responses": {
                    "200": {
                        "description": "A list of tournaments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tournament"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Create a new tournament",
                "operationId": "d6053f0972883878f37e936e0316a92d",
                "parameters": [
                    {
                        "name": "X-CSRF-TOKEN",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TournamentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tournament created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tournament"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tournaments/{id}": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get tournament by ID",
                "operationId": "6fc0b01dfb2e637254d85be733c11867",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tournament"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tournament not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Update a tournament",
                "operationId": "8005525d118e8b5faf9780fe28777f31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-CSRF-TOKEN",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TournamentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tournament updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tournament"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tournament not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Delete a tournament",
                "operationId": "a34335cdf270ec1cbbe62c0062364dc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tournament deleted"
                    },
                    "404": {
                        "description": "Tournament not found"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securepassword"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "roles": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User retrieved successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "roles": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TournamentRequest": {
                "required": [
                    "tournament_name",
                    "vendor_name",
                    "description",
                    "image",
                    "current_stage",
                    "current_session",
                    "track_number",
                    "champ_number",
                    "bto_number",
                    "status"
                ],
                "properties": {
                    "tournament_name": {
                        "description": "The name of the tournament",
                        "type": "string"
                    },
                    "vendor_name": {
                        "description": "The name of the vendor",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the tournament",
                        "type": "string"
                    },
                    "image": {
                        "description": "Image URL of the tournament",
                        "type": "string"
                    },
                    "current_stage": {
                        "description": "The current stage of the tournament",
                        "type": "integer"
                    },
                    "current_session": {
                        "description": "The current session of the tournament",
                        "type": "integer"
                    },
                    "track_number": {
                        "description": "The track number",
                        "type": "integer"
                    },
                    "champ_number": {
                        "description": "The champion number",
                        "type": "integer"
                    },
                    "bto_number": {
                        "description": "The back-to-office number",
                        "type": "integer"
                    },
                    "status": {
                        "description": "The status of the tournament",
                        "type": "string"
                    },
                    "created_by": {
                        "description": "ID of the user who created the tournament",
                        "type": "string",
                        "nullable": true
                    },
                    "updated_by": {
                        "description": "ID of the user who last updated the tournament",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Tournament": {
                "properties": {
                    "id": {
                        "description": "The UUID of the tournament",
                        "type": "string"
                    },
                    "tournament_name": {
                        "description": "The name of the tournament",
                        "type": "string"
                    },
                    "vendor_name": {
                        "description": "The name of the vendor",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the tournament",
                        "type": "string"
                    },
                    "image": {
                        "description": "Image URL of the tournament",
                        "type": "string"
                    },
                    "current_stage": {
                        "description": "The current stage of the tournament",
                        "type": "integer"
                    },
                    "current_session": {
                        "description": "The current session of the tournament",
                        "type": "integer"
                    },
                    "track_number": {
                        "description": "The track number",
                        "type": "integer"
                    },
                    "champ_number": {
                        "description": "The champion number",
                        "type": "integer"
                    },
                    "bto_number": {
                        "description": "The back-to-office number",
                        "type": "integer"
                    },
                    "status": {
                        "description": "The status of the tournament",
                        "type": "string"
                    },
                    "created_by": {
                        "description": "ID of the user who created the tournament",
                        "type": "string",
                        "nullable": true
                    },
                    "updated_by": {
                        "description": "ID of the user who last updated the tournament",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tournaments",
            "description": "Operations about tournaments"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}